+++ src/Minimizer.hpp
@@ -1,0 +1,8 @@
+#include "Directive.hpp"
+#include <string>
+#include <vector>
+#include <variant>
+
+using MinimizeCtx = std::vector<std::variant<std::string, Directive>>;
+
+std::string minimizeToStr(MinimizeCtx& mCtx);