+++ test/Default/Minimize/CorrectOutput/Test.cc
@@ -1,5 +1,3 @@
-#include "Export.hpp"
-#ifdef CPP_MODULES
 module;
 #if __has_include(<iostream>)
 #include <iostream>
@@ -7,21 +5,7 @@ module;
 #ifdef COND1
 #include "RandomHeader.hpp"
 #endif
-#ifdef COND2
-#define a
-#endif
 export module Test;
-#else
-#if __has_include(<iostream>)
-#include <iostream>
-#endif
-#ifdef COND1
-#include "RandomHeader.hpp"
-#endif
-#ifdef COND2
-#define a
-#endif
-#endif
 
 #ifdef COND 
 #endif 
@@ -40,3 +24,5 @@ export module Test;
 // Do something else
 #endif
 #endif
+#define A
+#undef A