+++ test/Default/Minimize/Input/Test.cc
@@ -1,20 +1,10 @@
-module;
+#ifdef COND 
+#endif 
 #if __has_include(<iostream>)
 #include <iostream>
 #endif
 #ifdef COND1
 #include "RandomHeader.hpp"
-#endif
-#ifdef COND2
-#define a
-#endif
-export module Test;
-
-#ifdef COND 
-#endif 
-#if __has_include(<iostream>)
-#endif
-#ifdef COND1
 #endif
 #ifdef COND2
 #define a
@@ -27,3 +17,5 @@ export module Test;
 // Do something else
 #endif
 #endif
+#define A
+#undef A