+++ test/Transitional/BackCompatHdrs/CorrectOutput/Test.cc
@@ -1,6 +1,8 @@
-#include "Export.hpp"
 #ifdef CPP_MODULES
 module;
+#endif
+#include "Export.hpp"
+#ifdef CPP_MODULES
 module Test;
 #else
 #include "Test.cppm"