+++ test/Transitional/IgnoredHdr/CorrectOutput/IgnoredHdr.cc
@@ -1,0 +1,5 @@
+#include "IgnoredHdr.hpp"
+#ifdef CPP_MODULES
+#else
+#endif
+