+++ test/Transitional/IgnoredHdr/CorrectOutput/IncludeIgnoredHdr.cc
@@ -1,9 +1,10 @@
-#include "Export.hpp"
 #ifdef CPP_MODULES
 module;
+#endif
+#include "Export.hpp"
 #include "IgnoredHdr.hpp"
+#ifdef CPP_MODULES
 export module IncludeIgnoredHdr;
 #else
-#include "IgnoredHdr.hpp"
 #endif