+++ test/Transitional/IgnoredHeader/CorrectOutput/IgnoredHdr.cc
@@ -1,6 +1,0 @@
-#ifdef CPP_MODULES
-#include "IgnoredHdr.hpp"
-#else
-#include "IgnoredHdr.hpp"
-#endif
-