+++ test/Transitional/Minimize/CorrectOutput/Test.cc
@@ -1,10 +1,23 @@
-#ifdef COND 
-#endif 
+#ifdef CPP_MODULES
+module;
+#endif
+#include "Export.hpp"
 #if __has_include(<iostream>)
 #include <iostream>
 #endif
 #ifdef COND1
 #include "RandomHeader.hpp"
+#endif
+#ifdef CPP_MODULES
+export module Test;
+#else
+#endif
+
+#ifdef COND 
+#endif 
+#if __has_include(<iostream>)
+#endif
+#ifdef COND1
 #endif
 #ifdef COND2
 #define a
@@ -17,3 +30,5 @@
 // Do something else
 #endif
 #endif
+#define A
+#undef A